+++ resources/attributes/default.rb
@@ -48,7 +48,7 @@ default['redborder']['services']['redborder-monitor'] = true
 default['redborder']['services']['redborder-scanner'] = true
 default['redborder']['services']['snmp'] = true
 default['redborder']['services']['f2k'] = true
-default['redborder']['services']['logstash'] = true
+default['redborder']['services']['logstash'] = false
 default['redborder']['services']['pmacct'] = true
 default['redborder']['services']['rsyslog'] = true
 default['redborder']['services']['redborder-nmsp'] = true
@@ -57,6 +57,7 @@ default['redborder']['services']['n2klocd'] = true
 default['redborder']['services']['radiusd'] = false
 default['redborder']['services']['k2http'] = true
 default['redborder']['services']['redborder-exporter'] = true
+default['redborder']['services']['clamav'] = true
 
 default['redborder']['systemdservices']['chef-client'] = ['chef-client']
 default['redborder']['systemdservices']['kafka'] = ['kafka']