+++ resources/recipes/configure.rb
@@ -99,7 +99,8 @@ logstash_config 'Configure logstash' do
   namespaces node.run_state['namespaces']
   vault_nodes node.run_state['sensors_info_all']['vault-sensor']
   device_nodes node.run_state['sensors_info_all']['device-sensor']
-  if proxy_services['logstash']
+  logstash_pipelines node.default['pipelines']
+  if !logstash_pipelines.nil? && !logstash_pipelines.empty?
     action [:add]
   else
     action [:remove]
@@ -167,6 +168,10 @@ end
 
 rbcgroup_config 'Configure cgroups' do
   action :add
+end
+
+rb_clamav_config 'Configure ClamAV' do
+  action(proxy_services['clamav'] ? :add : :remove)
 end
 
 # MOTD